Goliath walking pattern
Goliath now steps in a two at a time pattern. I fixed the feet script to be named better and more clearly (it's now UnderbodyMFoot), and also it can be scaled for any amount of feet and any walking pattern.

IN THE FUTURE, I will be doing more commits, probably after every small random complete feature, like this.